Drop Pipeline DSL; Only support Freestyle jobs
Not everyone can run Pipeline DSL, but every one can run freestyle jobs.
We only want to support one way to create jobs going forward.
