--- src/main/resources/cli/jenkins/pipeline.groovy.template
@@ -1,19 +1,0 @@
-pipeline {
-    agent %s
-    options {
-        timeout(time: 1, unit: 'HOURS')
-        disableConcurrentBuilds(abortPrevious: true)
-    }
-    %s
-    stages {
-%s
-%s
-%s
-    }
-    post {
-        always {
-            archiveArtifacts artifacts: '.moderne/build/**/build.log'
-            %s
-        }
-    }
-}
