--- src/main/resources/cli/jenkins/pipeline_credentials.groovy.template
@@ -1,3 +1,0 @@
-                withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: '%s', usernameVariable: 'ARTIFACTS_PUBLISH_CRED_USR', passwordVariable: 'ARTIFACTS_PUBLISH_CRED_PWD']]) {
-                    %s
-                }
