--- src/main/resources/cli/jenkins/pipeline_download.groovy.template
@@ -1,2 +1,0 @@
-                sh "curl --request GET %s --fail -o mod"
-                sh "chmod 755 mod"
