--- src/main/resources/cli/jenkins/pipeline_download_creds.groovy.template
@@ -1,4 +1,0 @@
-                withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: '%s', usernameVariable: 'CLI_DOWNLOAD_CRED_USR', passwordVariable: 'CLI_DOWNLOAD_CRED_PWD']]) {
-                    sh "curl --user ${CLI_DOWNLOAD_CRED_USR}:${CLI_DOWNLOAD_CRED_PWD} --request GET %s --fail -o mod"
-                    sh "chmod 755 mod"
-                }
