--- src/main/resources/cli/jenkins/pipeline_download_windows.groovy.template
@@ -1,4 +1,0 @@
-                powershell '''
-                $wc = New-Object System.Net.WebClient
-                $wc.DownloadFile("%s", "mod.exe")
-                '''
