--- src/main/resources/cli/jenkins/pipeline_environment_validate.groovy.template
@@ -1,4 +1,0 @@
-environment {
-        MODERNE_TOKEN = credentials('modToken')
-        SCM_TOKEN = credentials('%s')
-    }
