--- src/main/resources/cli/jenkins/pipeline_flow_definition.xml.template
@@ -1,41 +1,0 @@
-<flow-definition plugin="workflow-job@%s">
-    <actions>
-        <org.jenkinsci.plugins.pipeline.modeldefinition.actions.DeclarativeJobAction plugin="pipeline-model-definition@%s"/>
-        <org.jenkinsci.plugins.pipeline.modeldefinition.actions.DeclarativeJobPropertyTrackerAction plugin="pipeline-model-definition@%s">
-            <jobProperties/>
-            <triggers>
-                <string>hudson.triggers.TimerTrigger</string>
-            </triggers>
-            <parameters/>
-            <options/>
-        </org.jenkinsci.plugins.pipeline.modeldefinition.actions.DeclarativeJobPropertyTrackerAction>
-    </actions>
-    <description/>
-    <keepDependencies>false</keepDependencies>
-    <properties>
-        <jenkins.model.BuildDiscarderProperty>
-            <strategy class="hudson.tasks.LogRotator">
-                <daysToKeep>-1</daysToKeep>
-                <numToKeep>3</numToKeep>
-                <artifactDaysToKeep>-1</artifactDaysToKeep>
-                <artifactNumToKeep>3</artifactNumToKeep>
-            </strategy>
-        </jenkins.model.BuildDiscarderProperty>
-        <org.jenkinsci.plugins.workflow.job.properties.PipelineTriggersJobProperty>
-            <triggers>
-                <hudson.triggers.TimerTrigger>
-                    <spec>%s</spec>
-                </hudson.triggers.TimerTrigger>
-            </triggers>
-        </org.jenkinsci.plugins.workflow.job.properties.PipelineTriggersJobProperty>
-        %s
-    </properties>
-    <definition class="org.jenkinsci.plugins.workflow.cps.CpsFlowDefinition" plugin="workflow-cps@%s">
-        <script><![CDATA[
-%s
-        ]]></script>
-        <sandbox>true</sandbox>
-    </definition>
-    <triggers/>
-    <disabled>false</disabled>
-</flow-definition>
