--- src/main/resources/cli/jenkins/pipeline_maven_settings.groovy.template
@@ -1,3 +1,0 @@
-                configFileProvider([configFile(fileId: '%s', variable: 'MODERNE_MVN_SETTINGS_XML')]) {
-                    %s
-                }
