--- src/main/resources/cli/jenkins/pipeline_moderne_creds.groovy.template
@@ -1,3 +1,0 @@
-                withCredentials([string(credentialsId: '%s', variable: 'MODERNE_TOKEN')]) {
-                    %s
-                }
