--- src/main/resources/cli/jenkins/pipeline_stage_checkout.groovy.template
@@ -1,5 +1,0 @@
-        stage('Checkout') {
-            steps {
-                git (url: '%s', branch: '%s', credentialsId: '%s')
-            }
-        }
