--- src/main/resources/cli/jenkins/pipeline_stage_download.groovy.template
@@ -1,5 +1,0 @@
-        stage('Download CLI') {
-            steps {
-%s
-            }
-        }
