--- src/main/resources/cli/jenkins/pipeline_stage_publish.groovy.template
@@ -1,8 +1,0 @@
-        stage('Publish') {
-%s
-            steps {
-%s%s%s%s%s
-%s
-%s
-            }
-        }
