--- src/main/resources/cli/jenkins/pipeline_stage_validate.groovy.template
@@ -1,22 +1,0 @@
-        stage('Validate') {
-            %s
-            steps {
-                script {
-                    currentBuild.displayName = "\${params.buildName}"
-                }
-                %s
-                script {
-                    if (params.patchDownloadUrl) {
-                        withEnv(["PATCH_URL=\${params.patchDownloadUrl}"]) {
-                           println 'fetching patch'
-                           sh 'curl -o patch.diff --request GET --url \$PATCH_URL --header "Authorization: Bearer \$MODERNE_TOKEN" --header "x-moderne-scmtoken: \$SCM_TOKEN"'
-                        }
-                        println 'applying patch'
-                        sh 'git apply patch.diff'
-                    } else {
-                        println 'building without patch'
-                    }
-                }
-                %s
-            }
-        }
