--- src/main/resources/cli/jenkins/pipeline_tools.groovy.template
@@ -1,3 +1,0 @@
-            tools {
-                %s
-            }
