+++ src/main/resources/cli/jenkins/validate_parameters.xml.template
