+++ src/test/java/io/moderne/connect/commands/JenkinsTest.java
@@ -15,177 +15,397 @@
  */
 package io.moderne.connect.commands;
 
-import org.intellij.lang.annotations.Language;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import kong.unirest.HttpResponse;
+import kong.unirest.Unirest;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.containers.wait.strategy.Wait;
+import org.testcontainers.images.builder.ImageFromDockerfile;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+import picocli.CommandLine;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
+import static org.junit.jupiter.api.Assertions.*;
 
+@Testcontainers
 class JenkinsTest {
-    Jenkins jenkins = new Jenkins();
+    private final CommandLine cmd = new CommandLine(new Connect());
+    private static final String ARTIFACTORY_URL = "https://artifactory.moderne.ninja/artifactory/moderne-ingest";
+    private static final String ARTIFACT_CREDS = "artifactCreds";
+    private static final String GIT_CREDS = "myGitCreds";
+    private static final String MAVEN_SETTINGS = "maven_settings";
+    private static final String MODERNE_URL = "https://app.moderne.io";
+    private static final String MODERNE_TOKEN = "modToken";
+    private static final String JENKINS_TESTING_USER = "admin";
+    private static final String JENKINS_TESTING_PWD = "jenkins123";
+    private static final String AST_PUBLISH_USERNAME = "admin";
+    private static final String AST_PUBLISH_PASSWORD = "blah";
+    private static String jenkinsHost;
+    private String apiToken;
 
-    @Nested
-    class DownloadStage {
-        @Test
-        void download() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            jenkins.downloadCLI = true;
-            jenkins.platform = "linux";
-            jenkins.cliVersion = "v1.0.3"; // TODO do we really want this hardcoded into the command option as a default value?
+    @SuppressWarnings("resource")
+    @Container
+    private final GenericContainer<?> jenkinsContainer = new GenericContainer<>(
+            new ImageFromDockerfile()
+                    .withDockerfile(new File("src/test/jenkins/Dockerfile").toPath())
+                    .withFileFromFile("casc.yaml", new File("src/test/jenkins/casc.yaml")))
+            .withReuse(true)
+            .withExposedPorts(8080)
+            .withEnv("JENKINS_ADMIN_ID", JENKINS_TESTING_USER)
+            .withEnv("JENKINS_ADMIN_PASSWORD", JENKINS_TESTING_PWD)
+            .withEnv("JENKINS_AST_PUBLISH_USERNAME", AST_PUBLISH_USERNAME)
+            .withEnv("JENKINS_AST_PUBLISH_PASSWORD", AST_PUBLISH_PASSWORD)
+            .withEnv("JENKINS_GIT_USERNAME", "")
+            .withEnv("JENKINS_GIT_PASSWORD", "")
+            .waitingFor(Wait.forLogMessage(".*Jenkins is fully up and running.*\\n", 1));
 
-            assertDownloadSteps("""
-                    sh "curl --request GET https://pkgs.dev.azure.com/moderneinc/moderne_public/_packaging/moderne/maven/v1/io/moderne/moderne-cli-linux/v1.0.3/moderne-cli-linux-v1.0.3 --fail -o mod"
-                    sh "chmod 755 mod"
-                    """);
-        }
+    @BeforeEach
+    void setUp() {
+        jenkinsHost = "http://" + jenkinsContainer.getHost() + ":" + jenkinsContainer.getFirstMappedPort();
+        apiToken = createApiToken();
+    }
 
-        @Test
-        void customUrl() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            jenkins.downloadCLIUrl = "https://acme.com/moderne-cli";
-            assertDownloadSteps("""
-                    sh "curl --request GET https://acme.com/moderne-cli --fail -o mod"
-                    sh "chmod 755 mod"
-                    """);
-        }
-
-        @Test
-        void credentialsAndCustomUrl() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            jenkins.downloadCLIUrl = "https://acme.com/moderne-cli";
-            jenkins.downloadCLICreds = "downloadCreds";
-            assertDownloadSteps("""
-                    withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: 'downloadCreds', usernameVariable: 'CLI_DOWNLOAD_CRED_USR', passwordVariable: 'CLI_DOWNLOAD_CRED_PWD']]) {
-                        sh "curl --user ${CLI_DOWNLOAD_CRED_USR}:${CLI_DOWNLOAD_CRED_PWD} --request GET https://acme.com/moderne-cli --fail -o mod"
-                        sh "chmod 755 mod"
-                    }
-                    """);
-        }
-
-        void assertDownloadSteps(@Language("groovy") String steps) {
-            assertThat(jenkins.createStageDownload())
-                    .isEqualToIgnoringWhitespace("stage('Download CLI') { steps { %s } }".formatted(steps));
+    private static String createApiToken() {
+        // Create the API token, which appears not to be supported other than through the UI_main/api
+        HttpResponse<String> response = Unirest.post(jenkinsHost + "/me/descriptorByName/jenkins.security.ApiTokenProperty/generateNewToken")
+                .basicAuth(JENKINS_TESTING_USER, JENKINS_TESTING_PWD)
+                .header("Accept", "application/json")
+                .header("Content-Type", "application/x-www-form-urlencoded")
+                .header(Jenkins.JENKINS_CRUMB_HEADER, Jenkins.generateCrumb(jenkinsHost, JENKINS_TESTING_USER, JENKINS_TESTING_PWD))
+                .body("newTokenName=cli")
+                .asString();
+        assertTrue(response.isSuccess(), "Failed to create API token: " + response.getStatus() + " " + response.getStatusText());
+        try {
+            return new ObjectMapper()
+                    .readTree(response.getBody())
+                    .get("data")
+                    .get("tokenValue")
+                    .asText();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
         }
     }
 
+    @Test
+    void submitJobs() throws Exception {
+        int result = cmd.execute("jenkins",
+                "--fromCsv", new File("src/test/csv/repos.csv").getAbsolutePath(),
+                "--controllerUrl", jenkinsHost,
+                "--jenkinsUser", JENKINS_TESTING_USER,
+                "--apiToken", apiToken,
+                "--publishCredsId", ARTIFACT_CREDS,
+                "--gitCredsId", GIT_CREDS,
+                "--publishUrl", ARTIFACTORY_URL,
+                "--gradlePluginVersion", "2.2.2",
+                "--mirrorUrl", "http://artifactory.moderne.internal/artifactory/moderne-cache-3",
+                "--mvnPluginVersion", "2.4.2",
+                "--workspaceCleanup",
+                "--verbose");
+        assertEquals(0, result);
+
+        await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/api/json")
+                .asString().isSuccess()));
+
+        HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/config.xml").asString();
+        assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+        String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/config.xml").toPath()));
+        assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+    }
+
+    @Test
+    void submitJobsWithPassword() throws Exception {
+        int result = cmd.execute("jenkins",
+                "--fromCsv", new File("src/test/csv/repos.csv").getAbsolutePath(),
+                "--controllerUrl", jenkinsHost,
+                "--jenkinsUser", JENKINS_TESTING_USER,
+                "--jenkinsPwd", JENKINS_TESTING_PWD,
+                "--publishCredsId", ARTIFACT_CREDS,
+                "--gitCredsId", GIT_CREDS,
+                "--publishUrl", ARTIFACTORY_URL,
+                "--gradlePluginVersion", "2.2.2",
+                "--mirrorUrl", "http://artifactory.moderne.internal/artifactory/moderne-cache-3",
+                "--mvnPluginVersion", "2.4.2",
+                "--workspaceCleanup");
+        assertEquals(0, result);
+
+        await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/api/json")
+                .asString().isSuccess()));
+
+        HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/config.xml").asString();
+        assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+        String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/config.xml").toPath()));
+        assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+    }
+
+    @Test
+    void submitMultipleJobs() throws Exception {
+        int result = cmd.execute("jenkins",
+                "--fromCsv", new File("src/test/csv/jenkins-repos.csv").getAbsolutePath(),
+                "--controllerUrl", jenkinsHost,
+                "--jenkinsUser", JENKINS_TESTING_USER,
+                "--apiToken", apiToken,
+                "--publishCredsId", ARTIFACT_CREDS,
+                "--gitCredsId", GIT_CREDS,
+                "--publishUrl", ARTIFACTORY_URL,
+                "--workspaceCleanup");
+        assertEquals(0, result);
+
+        await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-java-migration_main/api/json").asString().isSuccess()));
+
+        HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-java-migration_main/config.xml").asString();
+        assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+        String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/rewrite-java-migration-config.xml").toPath()));
+        assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+    }
+
+    @Test
+    void submitJobTwice() throws Exception {
+        int result = cmd.execute("jenkins",
+                "--fromCsv", new File("src/test/csv/repos.csv").getAbsolutePath(),
+                "--controllerUrl", jenkinsHost,
+                "--jenkinsUser", JENKINS_TESTING_USER,
+                "--apiToken", apiToken,
+                "--publishCredsId", ARTIFACT_CREDS,
+                "--gitCredsId", GIT_CREDS,
+                "--publishUrl", ARTIFACTORY_URL,
+                "--gradlePluginVersion", "2.2.2",
+                "--mirrorUrl", "http://artifactory.moderne.internal/artifactory/moderne-cache-3",
+                "--mvnPluginVersion", "2.4.2",
+                "--workspaceCleanup");
+        assertEquals(0, result);
+
+        await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/api/json").asString().isSuccess()));
+
+        HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/config.xml").asString();
+        assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+        String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/config.xml").toPath()));
+        assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+
+        // Submit the same job again, to very that it doesn't fail
+        result = cmd.execute("jenkins",
+                "--fromCsv", new File("src/test/csv/repos.csv").getAbsolutePath(),
+                "--controllerUrl", jenkinsHost,
+                "--jenkinsUser", JENKINS_TESTING_USER,
+                "--apiToken", apiToken,
+                "--publishCredsId", ARTIFACT_CREDS,
+                "--gitCredsId", GIT_CREDS,
+                "--publishUrl", ARTIFACTORY_URL,
+                "--gradlePluginVersion", "2.2.2",
+                "--mirrorUrl", "http://artifactory.moderne.internal/artifactory/moderne-cache-3",
+                "--mvnPluginVersion", "2.4.2",
+                "--workspaceCleanup");
+        assertEquals(0, result);
+
+        await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/api/json").asString().isSuccess()));
+
+        response = Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/config.xml").asString();
+        assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+        assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+    }
+
+    @Test
+    void submitJobAndRemove() throws Exception {
+        int result = cmd.execute("jenkins",
+                "--fromCsv", new File("src/test/csv/repos.csv").getAbsolutePath(),
+                "--controllerUrl", jenkinsHost,
+                "--jenkinsUser", JENKINS_TESTING_USER,
+                "--apiToken", apiToken,
+                "--publishCredsId", ARTIFACT_CREDS,
+                "--gitCredsId", GIT_CREDS,
+                "--publishUrl", ARTIFACTORY_URL,
+                "--gradlePluginVersion", "2.2.2",
+                "--mirrorUrl", "http://artifactory.moderne.internal/artifactory/moderne-cache-3",
+                "--mvnPluginVersion", "2.4.2",
+                "--workspaceCleanup");
+        assertEquals(0, result);
+
+        await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/api/json").asString().isSuccess()));
+
+        HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/config.xml").asString();
+        assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+        String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/config.xml").toPath()));
+        assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+
+        // Now delete the job by marking it as skipped
+        result = cmd.execute("jenkins",
+                "--fromCsv", new File("src/test/csv/jenkins-skipped.csv").getAbsolutePath(),
+                "--controllerUrl", jenkinsHost,
+                "--jenkinsUser", JENKINS_TESTING_USER,
+                "--apiToken", apiToken,
+                "--publishCredsId", ARTIFACT_CREDS,
+                "--gitCredsId", GIT_CREDS,
+                "--publishUrl", ARTIFACTORY_URL,
+                "--gradlePluginVersion", "2.2.2",
+                "--mirrorUrl", "http://artifactory.moderne.internal/artifactory/moderne-cache-3",
+                "--mvnPluginVersion", "2.4.2",
+                "--workspaceCleanup",
+                "--deleteSkipped=true");
+        assertEquals(0, result);
+
+        await().untilAsserted(() -> assertFalse(Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/api/json").asString().isSuccess()));
+    }
+
+    @Test
+    void submitJobAgentWithXMLEntities() throws Exception {
+        int result = cmd.execute("jenkins",
+                "--fromCsv", new File("src/test/csv/repos.csv").getAbsolutePath(),
+                "--controllerUrl", jenkinsHost,
+                "--agent", "os=windows && !reserved",
+                "--jenkinsUser", JENKINS_TESTING_USER,
+                "--apiToken", apiToken,
+                "--publishCredsId", ARTIFACT_CREDS,
+                "--gitCredsId", GIT_CREDS,
+                "--publishUrl", ARTIFACTORY_URL,
+                "--workspaceCleanup");
+        assertEquals(0, result);
+
+        await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/api/json")
+                .asString().isSuccess()));
+
+        HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/config.xml").asString();
+        assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+        String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/config-agent.xml").toPath()));
+        assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+    }
+
+    @Test
+    void submitJobsNoCleanup() throws Exception {
+        int result = cmd.execute("jenkins",
+                "--fromCsv", new File("src/test/csv/repos.csv").getAbsolutePath(),
+                "--controllerUrl", jenkinsHost,
+                "--jenkinsUser", JENKINS_TESTING_USER,
+                "--apiToken", apiToken,
+                "--publishCredsId", ARTIFACT_CREDS,
+                "--gitCredsId", GIT_CREDS,
+                "--publishUrl", ARTIFACTORY_URL,
+                "--verbose");
+        assertEquals(0, result);
+
+        await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/api/json")
+                .asString().isSuccess()));
+
+        HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/moderne-ingest/job/openrewrite_rewrite-spring_main/config.xml").asString();
+        assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+        String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/config-no-cleanup.xml").toPath()));
+        assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+    }
+
     @Nested
-    class PublishStage {
+    class FreestyleJobs {
         @Test
-        void withoutJava() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            assertPublishSteps("""
-                    sh 'mod build . --no-download'
-                    sh 'mod publish .'
-                    """);
+        void submitFreestyleJobs() throws Exception {
+            int result = cmd.execute("jenkins",
+                    "--fromCsv", new File("src/test/csv/jenkins-repos.csv").getAbsolutePath(),
+                    "--controllerUrl", jenkinsHost,
+                    "--jenkinsUser", JENKINS_TESTING_USER,
+                    "--apiToken", apiToken,
+                    "--publishCredsId", ARTIFACT_CREDS,
+                    "--gitCredsId", GIT_CREDS,
+                    "--publishUrl", ARTIFACTORY_URL,
+                    "--folder", "freestyle",
+                    "--downloadCLI",
+                    "--mavenSettingsConfigFileId", MAVEN_SETTINGS,
+                    "--moderneUrl=" + MODERNE_URL,
+                    "--moderneToken=" + MODERNE_TOKEN,
+                    "--workspaceCleanup",
+                    "--gradlePluginVersion", "2.2.2",
+                    "--mirrorUrl", "http://artifactory.moderne.internal/artifactory/moderne-cache-3",
+                    "--mvnPluginVersion", "2.4.2",
+                    "--verbose");
+            assertEquals(0, result);
+
+            await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/freestyle/job/openrewrite_rewrite-spring_main/api/json")
+                    .asString().isSuccess()));
+
+            HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/freestyle/job/openrewrite_rewrite-spring_main/config.xml").asString();
+            assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+            String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/config-freestyle-gradle.xml").toPath()));
+            assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+
+            HttpResponse<String> responseMaven = Unirest.get(jenkinsHost + "/job/freestyle/job/openrewrite_rewrite-maven-plugin_main/config.xml").asString();
+            assertTrue(responseMaven.isSuccess(), "Failed to get job config.xml: " + responseMaven.getStatusText());
+            String expectedJobMaven = new String(Files.readAllBytes(new File("src/test/jenkins/config-freestyle-maven.xml").toPath()));
+            assertThat(responseMaven.getBody()).isEqualToIgnoringWhitespace(expectedJobMaven);
+
+            await().untilAsserted(() -> assertFalse(Unirest.get(jenkinsHost + "/job/validate/job/openrewrite_rewrite-spring_main/api/json")
+                    .asString().isSuccess()));
+
         }
 
         @Test
-        void windows() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            jenkins.platform = "windows";
-            jenkins.publishUrl = "https://my.artifactory/moderne-ingest";
-            jenkins.publishCredsId = "artifactCreds";
-            assertPublishSteps("""
-                    withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: 'artifactCreds', usernameVariable: 'ARTIFACTS_PUBLISH_CRED_USR', passwordVariable: 'ARTIFACTS_PUBLISH_CRED_PWD']]) {
-                        powershell 'mod.exe config artifacts edit --user=$env:ARTIFACTS_PUBLISH_CRED_USR --password=$env:ARTIFACTS_PUBLISH_CRED_PWD https://my.artifactory/moderne-ingest'
-                    }
-                    powershell 'mod.exe build . --no-download'
-                    powershell 'mod.exe publish .'
-                    """);
+        void submitFreestyleJobsWithValidate() throws Exception {
+            int result = cmd.execute("jenkins",
+                    "--fromCsv", new File("src/test/csv/jenkins-repos.csv").getAbsolutePath(),
+                    "--controllerUrl", jenkinsHost,
+                    "--jenkinsUser", JENKINS_TESTING_USER,
+                    "--apiToken", apiToken,
+                    "--publishCredsId", ARTIFACT_CREDS,
+                    "--gitCredsId", GIT_CREDS,
+                    "--publishUrl", ARTIFACTORY_URL,
+                    "--folder", "freestyle",
+                    "--downloadCLI",
+                    "--mavenSettingsConfigFileId", MAVEN_SETTINGS,
+                    "--moderneUrl=" + MODERNE_URL,
+                    "--moderneToken=" + MODERNE_TOKEN,
+                    "--createValidateJobs",
+                    "--workspaceCleanup",
+                    "--gradlePluginVersion", "2.2.2",
+                    "--mirrorUrl", "http://artifactory.moderne.internal/artifactory/moderne-cache-3",
+                    "--mvnPluginVersion", "2.4.2",
+                    "--verbose");
+            assertEquals(0, result);
+
+            await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/validate/job/openrewrite_rewrite-spring_main/api/json")
+                    .asString().isSuccess()));
+
+            HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/validate/job/openrewrite_rewrite-spring_main/config.xml").asString();
+            assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+            String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/config-freestyle-gradle-validate.xml").toPath()));
+            assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
+
+            HttpResponse<String> responseMaven = Unirest.get(jenkinsHost + "/job/validate/job/openrewrite_rewrite-maven-plugin_main/config.xml").asString();
+            assertTrue(responseMaven.isSuccess(), "Failed to get job config.xml: " + responseMaven.getStatusText());
+            String expectedJobMaven = new String(Files.readAllBytes(new File("src/test/jenkins/config-freestyle-maven-validate.xml").toPath()));
+            assertThat(responseMaven.getBody()).isEqualToIgnoringWhitespace(expectedJobMaven);
         }
+
 
         @Test
-        void skipSsl() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            jenkins.publishUrl = "https://my.artifactory/moderne-ingest";
-            jenkins.publishCredsId = "artifactCreds";
-            jenkins.skipSSL = true;
-            assertPublishSteps("""
-                    withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: 'artifactCreds', usernameVariable: 'ARTIFACTS_PUBLISH_CRED_USR', passwordVariable: 'ARTIFACTS_PUBLISH_CRED_PWD']]) {
-                        sh 'mod config artifacts edit --skipSSL --user=${ARTIFACTS_PUBLISH_CRED_USR} --password=${ARTIFACTS_PUBLISH_CRED_PWD} https://my.artifactory/moderne-ingest'
-                    }
-                    sh 'mod build . --no-download'
-                    sh 'mod publish .'
-                    """);
+        void submitFreestyleJobsNoCleanup() throws Exception {
+            int result = cmd.execute("jenkins",
+                    "--fromCsv", new File("src/test/csv/jenkins-repos.csv").getAbsolutePath(),
+                    "--controllerUrl", jenkinsHost,
+                    "--jenkinsUser", JENKINS_TESTING_USER,
+                    "--apiToken", apiToken,
+                    "--publishCredsId", ARTIFACT_CREDS,
+                    "--gitCredsId", GIT_CREDS,
+                    "--publishUrl", ARTIFACTORY_URL,
+                    "--folder", "freestyle",
+                    "--downloadCLI",
+                    "--mavenSettingsConfigFileId", MAVEN_SETTINGS,
+                    "--moderneUrl=" + MODERNE_URL,
+                    "--moderneToken=" + MODERNE_TOKEN,
+                    "--verbose");
+
+            assertEquals(0, result);
+
+            await().untilAsserted(() -> assertTrue(Unirest.get(jenkinsHost + "/job/freestyle/job/openrewrite_rewrite-spring_main/api/json")
+                    .asString().isSuccess()));
+
+            HttpResponse<String> response = Unirest.get(jenkinsHost + "/job/freestyle/job/openrewrite_rewrite-spring_main/config.xml").asString();
+            assertTrue(response.isSuccess(), "Failed to get job config.xml: " + response.getStatusText());
+            String expectedJob = new String(Files.readAllBytes(new File("src/test/jenkins/config-freestyle-gradle-no-cleanup.xml").toPath()));
+            assertThat(response.getBody()).isEqualToIgnoringWhitespace(expectedJob);
         }
 
-        @Test
-        void modConfigModerne() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            jenkins.tenant = new Jenkins.Tenant();
-            jenkins.tenant.moderneUrl = "https://app.moderne.io";
-            jenkins.tenant.moderneToken = "modToken";
-            jenkins.publishUrl = "https://my.artifactory/moderne-ingest";
-            jenkins.publishCredsId = "artifactCreds";
-            assertPublishSteps("""
-                    withCredentials([string(credentialsId: 'modToken', variable: 'MODERNE_TOKEN')]) {
-                        sh 'mod config moderne edit --token=${MODERNE_TOKEN} https://app.moderne.io'
-                    }
-                    withCredentials([[$class: 'UsernamePasswordMultiBinding', credentialsId: 'artifactCreds', usernameVariable: 'ARTIFACTS_PUBLISH_CRED_USR', passwordVariable: 'ARTIFACTS_PUBLISH_CRED_PWD']]) {
-                        sh 'mod config artifacts edit --user=${ARTIFACTS_PUBLISH_CRED_USR} --password=${ARTIFACTS_PUBLISH_CRED_PWD} https://my.artifactory/moderne-ingest'
-                    }
-                    sh 'mod build . --no-download'
-                    sh 'mod publish .'
-                    """);
-        }
-
-        @Test
-        void submitJobWithMavenSettings() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            jenkins.mavenSettingsConfigFileId = "maven-ingest-settings-credentials";
-            assertPublishSteps("""
-                    sh 'mod config maven settings edit "\\${MODERNE_MVN_SETTINGS_XML}"'
-                    configFileProvider([configFile(fileId: 'maven-ingest-settings-credentials', variable: 'MODERNE_MVN_SETTINGS_XML')]) {
-                        sh 'mod build . --no-download'
-                    }
-                    sh 'mod publish .'
-                    """);
-        }
-
-        @Test
-        void submitJobWithMavenPluginVersion() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            jenkins.mvnPluginVersion = "5.0.2";
-            assertPublishSteps("""
-                    sh 'mod config maven plugin edit --version 5.0.2'
-                    sh 'mod build . --no-download'
-                    sh 'mod publish .'
-                    """);
-        }
-
-        @Test
-        void submitJobWithGradlePluginVersion() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            jenkins.gradlePluginVersion = "5.0.2";
-            assertPublishSteps("""
-                    sh 'mod config gradle plugin edit --version 5.0.2'
-                    sh 'mod build . --no-download'
-                    sh 'mod publish .'
-                    """);
-        }
-
-        @Test
-        void withGradleTool() {
-            jenkins.jobType = Jenkins.JobType.PIPELINE;
-            assertThat(jenkins.createStagePublish("", "Gradle7", "", ""))
-                    //language=groovy
-                    .isEqualToIgnoringWhitespace("""
-                            stage('Publish') {
-                               tools {
-                                  gradle 'Gradle7'
-                               }
-                               steps {
-                                  sh 'mod build . --no-download'
-                                  sh 'mod publish .'
-                               }
-                            }
-                            """
-                    );
-        }
-
-        void assertPublishSteps(@Language("groovy") String steps) {
-            assertThat(jenkins.createStagePublish("maven", "gradle", "", ""))
-                    .isEqualToIgnoringWhitespace("stage('Publish') { tools { maven 'maven' gradle 'gradle' } steps { %s } }".formatted(steps));
-        }
     }
 }
